+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # obs2nlm ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added a `--dry-run` command line option.
+  ([#12](https://github.com/davep/obs2nlm/pull/12))
+
 ## v1.1.2
 
 **Released: 2026-01-11**