+++ ChangeLog.md
@@ -1,8 +1,8 @@
 # obs2nlm ChangeLog
 
-## Unreleased
+## v1.2.0
 
-**Released: WiP**
+**Released: 2026-04-10**
 
 - Added a `--dry-run` command line option.
   ([#12](https://github.com/davep/obs2nlm/pull/12))